refactor: decompose_utils
